Mouse Player controls now working correctly.

+++ EntityEngine/MouseMotionMessage.h
@@ -5,10 +5,10 @@
 
 class MouseMotionMessage : public Message {
 private:
-	glm::ivec2 _cursorPosition;
+	glm::vec2 _cursorPosition;
 
 public:
-	MouseMotionMessage(const glm::ivec2& cursorPosition)
+	MouseMotionMessage(const glm::vec2& cursorPosition)
 		: Message("mousemotion"),
 		  _cursorPosition(cursorPosition)
 	{
@@ -16,7 +16,7 @@ public:
 
 	virtual ~MouseMotionMessage() = default;
 
-	const glm::ivec2& GetCursorPosition() const noexcept { return _cursorPosition; }
+	const glm::vec2& GetCursorPosition() const noexcept { return _cursorPosition; }
 };
 
 #endif //MOUSE_MOTION_MESSAGE_H_
